+++ R/RcppExports.R
@@ -82,6 +82,10 @@ bsplineD1__impl <- function(x, j, order, knots) {
     .Call('cpr_bsplineD1__impl', PACKAGE = 'cpr', x, j, order, knots)
 }
 
+bsplineD2__impl <- function(x, j, order, knots) {
+    .Call('cpr_bsplineD2__impl', PACKAGE = 'cpr', x, j, order, knots)
+}
+
 bbasis__impl <- function(x, iknots, bknots, k) {
     .Call('cpr_bbasis__impl', PACKAGE = 'cpr', x, iknots, bknots, k)
 }