+++ src/deboor.cpp
@@ -149,6 +149,54 @@ Rcpp::NumericVector bsplineD1__impl(arma::vec x, unsigned int j, unsigned int or
   return out;
 }
 
+//[[Rcpp::export]]
+Rcpp::NumericVector bsplineD2__impl(arma::vec x, unsigned int j, unsigned int order, arma::vec knots) { 
+
+  double a = 0.0;
+  double b = 0.0;
+  arma::vec A1;
+  arma::vec A2;
+  arma::vec A3;
+  arma::vec A4;
+  A1.zeros(x.n_elem);
+  A2.zeros(x.n_elem);
+  A3.zeros(x.n_elem);
+  A4.zeros(x.n_elem);
+
+  a = knots(j + order) - knots(j + 2);
+  if (a != 0.0) { 
+    bspline Aj(x, j + 2, order - 2, knots);
+    A1 = -1.0 * Aj.get_Bj() / a;
+  }
+
+  a = knots(j + order - 1) - knots(j + 1);
+  if (a != 0.0) { 
+    bspline Bj(x, j + 1, order - 2, knots);
+    A2 = Bj.get_Bj() / a;
+  }
+  
+  a = knots(j + order - 2) - knots(j);
+  if (a != 0.0) {
+    bspline Cj(x, j, order - 2, knots);
+    A3 = Cj.get_Bj() / a;
+  }
+
+  a = knots(j + order) - knots(j + 1);
+  if (a != 0.0) { 
+    a = -1.0 / a;
+  }
+
+  b = knots(j + order - 1) - knots(j);
+  if (b != 0.0) { 
+    b = 1.0 / b;
+  }
+
+  A4 = double(order - 1) * double(order - 2) * (a * (A1 + A2) + b * (A3 - A2));
+
+  Rcpp::NumericVector out = arma2vec(A4); 
+  return out;
+}
+
 // [[Rcpp::export]]
 Rcpp::NumericMatrix bbasis__impl(arma::vec x, arma::vec iknots, arma::vec bknots, unsigned int k) {